django_project/requirements.txt
asgiref==3.5.0
asttokens==2.0.5
async-timeout==4.0.2
backcall==0.2.0
colorama==0.4.4
decorator==5.1.1
Deprecated==1.2.13
Django==4.0.3
django-debug-toolbar==3.2.4
django-reinstallation-app
djangorestframework==3.13.1
executing==0.8.3
ipython==8.2.0
jedi==0.18.1
matplotlib-inline==0.1.3
packaging==21.3
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.28
psycopg2==2.9.3
pure-eval==0.2.2
Pygments==2.11.2
pyparsing==3.0.7
pytz==2022.1
redis==4.2.0
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
traitlets==5.1.1
typing_extensions==4.1.1
tzdata==2022.1
wcwidth==0.2.5
wrapt==1.14.0
